pyproject.toml
[project]
name = "ai-coach"
version = "0.1.0"
description = "AI Coach"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.123.0",
    "garminconnect>=0.2.36",
    "google-cloud-secret-manager>=2.25.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
